120 77 evaluation 2

+++ KioChess/Engine/Interfaces/Config/IterativeDeepingConfiguration.cs
@@ -2,7 +2,7 @@
 
 public class IterativeDeepingConfiguration
 {
-    public int InitialDepth { get; set; }
-    public int DepthStep { get; set; }
+    public short InitialDepth { get; set; }
+    public short DepthStep { get; set; }
     public string[] Strategies { get; set; }
 }
